requirements.txt
streamlit
streamlit-authenticator
python-docx
blobfile
reportlab